fix(tktrex): handle video route event and send it to the API
