+++ packages/shared/src/extension/app.ts
@@ -125,7 +125,7 @@ function setupObserver(
 
       if (window?.document) {
         if (platformMatch.test(window.location.href)) {
-          if (oldHref !== window.location.href) {
+          if (window.location.href !== oldHref) {
             const newHref = window.location.href;
 
             appLog.debug(
@@ -136,28 +136,33 @@ function setupObserver(
 
             onLocationChange(oldHref, newHref);
 
-            oldHref = newHref;
-          }
+            // always call the route handler
+            const routeHandlerKey = handlersList.find((h) => {
+              const handler = handlers[h];
 
-          // always call the route handler
-          const routeHandlerKey = handlersList.find((h) => {
-            const handler = handlers[h];
+              if (handler.match.type === 'route') {
+                appLog.debug(
+                  'Matching route %O',
+                  handler.match,
+                  window.location.pathname
+                );
+                return window.location.pathname.match(handler.match.location);
+              }
+              return false;
+            });
 
-            if (handler.match.type === 'route') {
-              return window.location.pathname.match(handler.match.location);
+            if (routeHandlerKey) {
+              appLog.debug('Route handler key %s', routeHandlerKey);
+              const { handle, ...routeHandlerOpts } = handlers[
+                routeHandlerKey
+              ] as RouteObserverHandler;
+              handle(window.document.body, routeHandlerOpts, routeHandlerKey, {
+                ...config,
+                href: window.location.toString(),
+              } as any);
             }
-            return false;
-          });
 
-          if (routeHandlerKey) {
-            appLog.debug('Route handler key %s', routeHandlerKey);
-            const { handle, ...routeHandlerOpts } = handlers[
-              routeHandlerKey
-            ] as RouteObserverHandler;
-            handle(window.document.body, routeHandlerOpts, routeHandlerKey, {
-              ...config,
-              href: window.location.toString(),
-            } as any);
+            oldHref = newHref;
           }
         }
       }
@@ -171,7 +176,9 @@ function setupObserver(
 
   window.addEventListener('unload', () => {
     appLog.debug('Window unloading, disconnect the observer...');
-    observer.disconnect();
+    if (observer) {
+      observer.disconnect();
+    }
   });
 
   if (body) {
@@ -185,7 +192,7 @@ function setupObserver(
 
 /* Boot the application by giving a callback that
  * will be invoked after the handshake with API server.
- */
+ **/
 
 interface App {
   config: UserSettings;
